+++ src/wtp_mod.h
@@ -408,4 +408,6 @@ int __cdecl wtp_mod_probe_treaty_on(int faction1Id, int faction2Id, int treaty);
 int __cdecl wtp_mod_enemy_move(int vehicleId);
 int __cdecl wtp_mod_steal_energy(int baseId);
 int __cdecl wtp_mod_diplomacy_caption_say_fac_special(int dst, int src, int factionId);
+int wtp_mod_probe_success_rates_procure_research_data(int position, int morale, int risk, int baseId);
+int wtp_mod_probe_veh_skip(int vehicleId);